・コンテナ追加中（一旦コミット）。 ・ビルド設定（デバッグ機能有効化など）を追加。 ・ビルド設定・環境診断処理に、ビルド設定の表示を追加。また、直接表示せず、文字列を返すインターフェースに変更。

+++ include/gasha/shell_sort.h
@@ -179,7 +179,7 @@ std::size_t linkedListShellSort(T*& first, T*& last, GET_NEXT_FUNC get_next_func
 			node = get_next_func(*node);
 		}
 	}
-	const int h_max = size / 3;
+	const int h_max = static_cast<int>(size) / 3;
 	int h = 1;
 	while (h <= h_max)
 		h = 3 * h + 1;
